Added permission checks

+++ src/backend/django_src/calendars/views/calendars.py
@@ -1,4 +1,5 @@
 from rest_framework.permissions import IsAuthenticated
+from ..permissions import IsCalendarOwner
 from ..models.Calendar import Calendar
 from ..models.Member import Member
 from ..serializers import CalendarListSerializer, CalendarPUTSerializer
@@ -17,7 +18,7 @@ from rest_framework.views import APIView
 
 # EndPoint: /calendars/list/
 class CalendarList(APIView):
-    # permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated]
 
     def get(self, request):
         # Get all calendars that the user created(owner=user)
@@ -41,15 +42,19 @@ class CalendarList(APIView):
     
 # EndPoint: /calendars/<int:calendar_id>/
 class CalendarDetail(APIView):
-    # permission_classes = [IsAuthenticated] # Only the owner can view, edit, and delete the calendar
+    permission_classes = [IsAuthenticated, IsCalendarOwner]
 
     def get(self, request, calendar_id):
         calendar = get_object_or_404(Calendar, id=calendar_id)
+        self.check_object_permissions(request, calendar)
+
         serializer = CalendarListSerializer(calendar)
         return Response(serializer.data)
 
     def put(self, request, calendar_id):
         calendar = get_object_or_404(Calendar, id=calendar_id)
+        self.check_object_permissions(request, calendar)
+
         serializer = CalendarPUTSerializer(calendar, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -58,15 +63,18 @@ class CalendarDetail(APIView):
 
     def delete(self, request, calendar_id):
         calendar = get_object_or_404(Calendar, id=calendar_id)
+        self.check_object_permissions(request, calendar)
+
         calendar.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
     
 # EndPoint: /calendars/<int:calendar_id>/remindAll
 class CalendarRemind(APIView):
-    # permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsCalendarOwner]
     
     def post(self, request, calendar_id):
         calendar = get_object_or_404(Calendar, pk=calendar_id)
+        self.check_object_permissions(request, calendar)
         
         members = Member.objects.filter(calendar=calendar)
         owner_name = request.user.first_name
